Convert tests to PHPUnit
* Add development dependency on phpunit/phpunit ^6.0
* Update composer.lock for PHPUnit only
* Add test script to composer.json and use it in .travis.yml
* Put expected test results into a new directory
