+++ .travis.yml
@@ -15,7 +15,7 @@ install:
    - php box.phar build
 
 script:
-   - tests/test.php
+   - composer test
 
 deploy:
   provider: releases
