+++ tests/Results/BarInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace bar;
+
+interface BarInterface
+{
+    /**
+     * @param array $a
+     * @param int   $b
+     *
+     * @return float
+     */
+    public function baz(array $a, int $b): float;
+}
