+++ tests/Results/BazTrait.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace bar;
+
+trait BazTrait
+{
+    /**
+     * @param int $a
+     *
+     * @return \DateTime
+     */
+    protected function inTrait(int $a): \DateTime
+    {
+    }
+}
