+++ tests/Results/Foo.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace bar;
+
+/**
+ * Foo.
+ */
+class Foo
+{
+    public function bar($foo)
+    {
+    }
+
+    /**
+     * @param float $a
+     */
+    public function test(float $a)
+    {
+        $closure = function ($a, $c) {
+        };
+    }
+}
