+++ tests/Results/array_no_types.php
@@ -1,0 +1,10 @@
+<?php
+# Typed arrays get simply converted to array
+/**
+ * @param int[] $ints
+ * @param string[] $strings
+ * @param Foo[] $someClasses
+ * @return float[]
+ */
+function array_no_types(array $ints, array $strings, array $someClasses) : array{
+}
