+++ tests/Results/functions2.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace foo;
+
+/**
+ * @param string|null $a
+ *
+ * @return int
+ */
+function test(string $a = null): int
+{
+}
