+++ tests/Results/param_no_type.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * @param $noType
+ */
+function param_no_type($noType)
+{
+}
