+++ tests/Results/type_aliases_and_whitelisting.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @param integer $integer
+ * @param boolean $boolean
+ * @param real $real
+ * @param double $double
+ * @param callback $callback
+ * @param void $void
+ * @param mixed $mixed
+ * @param unknown $unknown
+ * @param Class $class
+ */
+function aliases(int $integer, bool $boolean, $real, float $double, callable $callback, $void, $mixed, $unkown, \Class $class) {
+}
